% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wesp_site-class.R
\name{as.wesp_site}
\alias{as.wesp_site}
\title{Convert a data.frame of WESP question responses to a \code{wesp_site}}
\usage{
as.wesp_site(data, site = NULL)
}
\arguments{
\item{data}{\code{data.frame} of questions and responses, ideally created by
reading in data with \code{\link[=load_wesp_data]{load_wesp_data()}}. Contains columns \code{q_no},
\code{response_no}, and one or more \code{site_[x]} columns.}

\item{site}{A number, or the name of a column in \code{data} indicating which site to
calculate, if more than one site in \code{data}. Defaults to the
first \code{site_[x]} column.}
}
\value{
An object of type \code{wesp_site}
}
\description{
Read a \code{data.frame} that has been read in by \code{\link[=load_wesp_data]{load_wesp_data()}}
and convert it into a \code{wesp_site} object, in preparation for
calculating the indicators. This function also calculates various
"derived" values that are used in indicator calculations.
}
