% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_jenks_score.R
\name{assign_jenks_score}
\alias{assign_jenks_score}
\title{Assign Jenks score based on calibration dataset}
\usage{
assign_jenks_score(
  ind_scores,
  calibration_scores,
  EcoP,
  report = TRUE,
  output_dir
)
}
\arguments{
\item{ind_scores}{A data.frame of indicator scores. The output of get_indicator_scores().}

\item{calibration_scores}{an internal dataset containing the calibration data for all sites. This can be updated by admin.}

\item{EcoP}{A character string specifying the region. Default = 'GD'}

\item{report}{A logical indicating whether to generate a report. Default = TRUE.}

\item{output_dir}{A character string specifying the directory to save the report. Default = NULL.}
}
\value{
A data.frame with indicator scores and jenks classification score (Low, Medium, High (L, M, H)).
}
\description{
Assign Jenks score based on calibration dataset
}
\examples{
\dontrun{
wesp <- fs::path("inst/input_data/wetFlat_20250427_single.csv")
wesp_data <- load_wesp_data(wesp)
site <- as.wesp_site(wesp_data)
site <- calc_indicators(site)
ind_scores <- get_indicator_scores(site)
out <- assign_jenks_score(ind_scores, calibration_scores, EcoP = "GD")
}
}
