% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-indicators.R
\name{calc_indicators}
\alias{calc_indicators}
\title{Calculate Indicators for WESP assessment site}
\usage{
calc_indicators(site)
}
\arguments{
\item{site}{An object of type \code{wesp_site}, as created
with \code{\link[=as.wesp_site]{as.wesp_site()}}.}
}
\value{
\code{wesp_site} object with indicators calculated.
}
\description{
Calculate all Function and Benefit indicators
for a site.
}
