% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_jenks_score.R
\name{calculate_jenks_score}
\alias{calculate_jenks_score}
\title{Calculate Jenks Breaks}
\usage{
calculate_jenks_score(
  wespdata,
  sites = NULL,
  out_dir,
  out_name = "wesp_scores.csv"
)
}
\arguments{
\item{wespdata}{A data frame containing the formatted wesp data. This is the output
of load_wesp_data()}

\item{sites}{A numeric with the number of sites if specific sites are to be calculated.
The default is NULL,
which will include all sites in the calculation}

\item{out_dir}{a character string or path to location where output file is to be saved}

\item{out_name}{a character string with the name of the output file.
Default values is "wesp_scores.csv"}
}
\value{
a dateframe with compiled raw, normalised and jenks break values for
each site and each ecosystem function or benefit.
}
\description{
Calculate Jenks Breaks
}
\examples{
\dontrun{
wespdata <- load_wesp_data(system.file("input_data/wetFlat_20240130.csv", package = "wespr"))
calculate_jenks_score(wespdata, out_dir = "temp",  out_name = "wesp_scores.csv")
}
}
