% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_multi_site.R
\name{calculate_multi_site}
\alias{calculate_multi_site}
\title{Calculate indicators for multiple sites}
\usage{
calculate_multi_site(wespdata, sites = NULL)
}
\arguments{
\item{wespdata}{A data frame containing the formatted wesp data. This is the output of load_wesp_data()}

\item{sites}{A numeric with the number of sites if specific sites are to be calculated. The default is NULL,
which will include all sites in the calculation}
}
\value{
a dataframe with raw values for each site and each ecosystem service or benefit.
}
\description{
Calculate indicators for multiple sites
}
\examples{
\dontrun{
wespdata <- load_wesp_data(system.file("input_data/wetFlat_20240130.csv", package = "wespr"))
calculate_multi_site(wespdata)
}
}
