% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{calibration_scores}
\alias{calibration_scores}
\title{WESP calibrated site values}
\format{
\subsection{\code{calibration_scores}}{

A data frame with 99 rows and 7 columns:
\describe{
\item{ecoprovince}{ecoprovince abbreviation}
\item{service}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
see developers
}
\usage{
calibration_scores
}
\description{
This dataset is created for the purpose of calibrating the WESP model.
The script to process this data is available : \code{data-raw/upload-calibration-data.R}.
}
\keyword{datasets}
