% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_indata.R
\name{check_indata}
\alias{check_indata}
\title{Check the input data for basic QA}
\usage{
check_indata(indata)
}
\arguments{
\item{indata}{A path to the input data}
}
\value{
messages to the console
}
\description{
Check the input data for basic QA
}
\examples{
\dontrun{
indata <- fs::path("inst/input_data/wetFlat_20250325.csv")
check_indata(indata)
}
}
