% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_rawdata.R
\name{combine_rawdata}
\alias{combine_rawdata}
\title{Combine all types of raw data}
\usage{
combine_rawdata(
  field_data,
  office_data,
  EcoP = "GD",
  write_subfiles = TRUE,
  out_dir = "temp",
  overwrite = FALSE
)
}
\arguments{
\item{field_data}{A path to the field data}

\item{office_data}{A path to the office data}

\item{EcoP}{A character string specifying the region. Default = 'GD'}

\item{write_subfiles}{A logical value specifying whether to write the subfiles. Default = TRUE}

\item{out_dir}{A character string specifying the output directory.}

\item{overwrite}{A logical value specifying whether to overwrite the output files. Default = FALSE}
}
\value{
A tibble with the combined data
}
\description{
Combine all types of raw data
}
\examples{
\dontrun{
combine_rawdata(
  field_data = fs::path("inst/input_data/raw", "field_survey123_edited.xls"),
  office_data = fs::path("inst/input_data/raw", "scripted_office.xlsx"),
  EcoP = "GD",
  write_subfiles = FALSE,
  out_dir = "temp",
  overwrite = TRUE
)
}
}
