% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{derive_values}
\alias{derive_values}
\title{Calculate derived values from question responses}
\usage{
derive_values(questions)
}
\arguments{
\item{questions}{an object resulting from running \code{\link[=record_values]{record_values()}}}
}
\value{
A \code{list} object containing derived values and their metadata.
}
\description{
Examples are \code{all_water}, \code{never_water}, etc.
}
