% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_rawinputs.R
\name{format_rawinputs}
\alias{format_rawinputs}
\title{Format raw inputs
This function cleans, formats and combines raw outputs from field and desktop analysis survey123 forms.}
\usage{
format_rawinputs(
  field_data,
  desktop_data,
  write_subfiles = TRUE,
  out_dir = "temp",
  overwrite = FALSE
)
}
\arguments{
\item{field_data}{A file path to survey123 csv file for field data collected}

\item{desktop_data}{A file path to survey123 csv file for Desktop data}

\item{write_subfiles}{A logical value specifying whether to write the subfiles. Default = TRUE}

\item{out_dir}{A character string specifying the output directory.}

\item{overwrite}{A logical value specifying whether to overwrite the output files. Default = FALSE}
}
\value{
dataframe in format for use in wespr
}
\description{
Format raw inputs
This function cleans, formats and combines raw outputs from field and desktop analysis survey123 forms.
}
\examples{
\dontrun{
ww <- format_rawinputs(
 field_data <- system.file(file.path('extdata','WESP_FIELDV1.csv'), package = "wespr"),
 desktop_data <- system.file(file.path('extdata','WESP_DESKTOPV1.csv'), package = "wespr"),
 write_subfiles = FALSE,
 out_dir = "input_data",
 overwrite = TRUE)
}
}
