% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{generate_ids}
\alias{generate_ids}
\title{Generate a key with names to be retained}
\usage{
generate_ids(wesp_data)
}
\arguments{
\item{wesp_data}{dataset which is output of load_wesp_data}
}
\value{
a \code{data.frame} of the responses object containing site names with site numbered key
}
\description{
Generate a key with names to be retained
}
