% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wesp-site-methods.R
\name{get_derived_values}
\alias{get_derived_values}
\title{Retrieve all derived values from a \code{wesp_site} object as a data.frame.}
\usage{
get_derived_values(site, ...)
}
\arguments{
\item{site}{A \code{wesp_site} object}

\item{...}{ignored for now}
}
\value{
A data.frame of responses
}
\description{
Retrieve all derived values from a \code{wesp_site} object as a data.frame.
}
