% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wesp-site-methods.R
\name{get_indicator_scores}
\alias{get_indicator_scores}
\title{Retrieve indicator scores from a \code{wesp_site} object as a data.frame.}
\usage{
get_indicator_scores(site, ...)
}
\arguments{
\item{site}{A \code{wesp_site} object, having been updated with \code{\link[=calc_indicators]{calc_indicators()}}}

\item{...}{ignored for now}
}
\value{
A data.frame of indicator scores
}
\description{
Retrieve indicator scores from a \code{wesp_site} object as a data.frame.
}
