% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_spatial_data.R
\name{import_spatial_data}
\alias{import_spatial_data}
\title{Download base spatial information layers to support desktop analysis for Wespr}
\usage{
import_spatial_data(aoi = NA, out_dir = NA)
}
\arguments{
\item{aoi}{An \code{sf} object (e.g. polygon) of designated wetland}

\item{out_dir}{A character string of filepath which points to output location.}
}
\value{
path to the output directory where files are written (invisibly).
}
\description{
Download base spatial information layers to support desktop analysis for Wespr
}
\examples{
\dontrun{
#import_spatial_data(aoi, out_dir = fs::path("./temp/spatial_layers"))
}


}
