% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{load_wesp_data}
\alias{load_wesp_data}
\title{Load the data containing the responses}
\usage{
load_wesp_data(path)
}
\arguments{
\item{path}{path to csv}
}
\value{
a \code{data.frame} of the responses
}
\description{
This file is usually an output from the R script at
https://github.com/BCWF-Wetlands/WESP_Calculator
}
