% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{record_values}
\alias{record_values}
\title{Validate and record responses into a standard object}
\usage{
record_values(data)
}
\arguments{
\item{data}{a \code{data.frame}, the output of \code{\link[=load_wesp_data]{load_wesp_data()}}}
}
\value{
a \code{list} object containing validated responses and question metadata
}
\description{
This currently is only implemented for a single site
}
