% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wespr-package.R
\docType{package}
\name{wespr-package}
\alias{wespr}
\alias{wespr-package}
\title{wespr: Calculate ecosystem function and benefit scores for wetlands}
\description{
Takes responses from a predetermined set of office and field-based questions, validates the responses, and calculates indicator scores for various aspects of wetland functions and benefits.
}
\author{
\strong{Maintainer}: Genevieve Perkins \email{gperkins@ninoxconsulting.ca}

Authors:
\itemize{
  \item Andy Teucher \email{andy.teucher@gmail.com} (\href{https://orcid.org/0000-0002-7840-692X}{ORCID})
}

Other contributors:
\itemize{
  \item BC Wildlife Federation [copyright holder]
}

}
\keyword{internal}
