% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{record_values}
\alias{record_values}
\title{Validate and record responses into a standard object
This is an internal function to generate wesp class object}
\usage{
record_values(data, site)
}
\arguments{
\item{data}{path to csv}

\item{site}{a number to show which site}
}
\value{
a \code{list} object containing validated responses and question metadata
}
\description{
Validate and record responses into a standard object
This is an internal function to generate wesp class object
}
