% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wespr_workflow.R
\name{wespr_workflow}
\alias{wespr_workflow}
\title{Title}
\usage{
wespr_workflow(desktop_data, field_data, out_dir, EcoP)
}
\arguments{
\item{desktop_data}{A file path to survey123 csv file for office data collected}

\item{field_data}{A file path to survey123 csv file for field data collected}

\item{out_dir}{A character string specifying the output directory.}

\item{EcoP}{A character string specifying the region. Default = 'GD'}
}
\value{
A tibble with the combined data
}
\description{
Title
}
\examples{
\dontrun{
wespr_workflow(
  field_data = fs::path("inst/input_data/raw", "field_survey123_edited.xls"),
  office_data = fs::path("inst/input_data/raw", "scripted_office.xlsx"),
  EcoP = "GD",
  write_subfiles = FALSE,
  out_dir = "temp",
  overwrite = TRUE)
}
}
